refactor(react): merge hoc's code into original component file
