+++ packages/react/src/Delay.tsx
@@ -1,4 +1,4 @@
-import { PropsWithChildren, createContext, useContext, useEffect, useState } from 'react'
+import { ComponentProps, ComponentType, PropsWithChildren, createContext, useContext, useEffect, useState } from 'react'
 import { PropsWithoutChildren } from './types'
 
 type DelayProps = PropsWithChildren<{
@@ -18,3 +18,22 @@ export const Delay = ({ ms, children }: DelayProps) => {
 }
 
 export const DelayContext = createContext<PropsWithoutChildren<DelayProps>>({ ms: 0 })
+
+// HOC
+export const withDelay = <TProps extends ComponentProps<ComponentType> = Record<string, never>>(
+  Component: ComponentType<TProps>,
+  delayProps?: PropsWithoutChildren<DelayProps>
+) => {
+  const Wrapped = (props: TProps) => (
+    <Delay {...delayProps}>
+      <Component {...props} />
+    </Delay>
+  )
+
+  if (process.env.NODE_ENV !== 'production') {
+    const name = Component.displayName || Component.name || 'Component'
+    Wrapped.displayName = `withDelay(${name})`
+  }
+
+  return Wrapped
+}
