+++ packages/react/src/Suspense.spec.tsx
@@ -1,6 +1,6 @@
 import { act, render, screen, waitFor } from '@testing-library/react'
 import { FALLBACK, MS_100, Suspend, TEXT } from './utils/toTest'
-import { Suspense } from '.'
+import { Suspense, withSuspense } from '.'
 
 describe('Suspense', () => {
   beforeEach(Suspend.reset)
@@ -55,5 +55,39 @@ describe('Suspense', () => {
     render(<Suspense.CSROnly fallback={FALLBACK}>{TEXT}</Suspense.CSROnly>)
     expect(screen.queryByText(FALLBACK)).not.toBeInTheDocument()
     expect(screen.queryByText(TEXT)).toBeInTheDocument()
+  })
+})
+
+const SuspendDuring100msToShowTEXTInSuspense = withSuspense(() => <Suspend during={MS_100} toShow={TEXT} />, {
+  fallback: FALLBACK,
+})
+const SuspendDuring100msToShowTEXTInSuspenseCSROnly = withSuspense.CSROnly(
+  () => <Suspend during={MS_100} toShow={TEXT} />,
+  { fallback: FALLBACK }
+)
+
+describe('withSuspense', () => {
+  beforeEach(Suspend.reset)
+
+  it('should wrap component by Suspense', async () => {
+    vi.useFakeTimers()
+    render(<SuspendDuring100msToShowTEXTInSuspense />)
+    expect(screen.queryByText(FALLBACK)).toBeInTheDocument()
+    expect(screen.queryByText(TEXT)).not.toBeInTheDocument()
+    vi.advanceTimersByTime(MS_100)
+    await waitFor(() => expect(screen.queryByText(FALLBACK)).not.toBeInTheDocument())
+    expect(screen.queryByText(TEXT)).toBeInTheDocument()
+    expect(screen.queryByText(FALLBACK)).not.toBeInTheDocument()
+  })
+
+  it('.CSROnly should wrap component by Suspense.CSROnly', async () => {
+    vi.useFakeTimers()
+    render(<SuspendDuring100msToShowTEXTInSuspenseCSROnly />)
+    expect(screen.queryByText(FALLBACK)).toBeInTheDocument()
+    expect(screen.queryByText(TEXT)).not.toBeInTheDocument()
+    vi.advanceTimersByTime(MS_100)
+    await waitFor(() => expect(screen.queryByText(FALLBACK)).not.toBeInTheDocument())
+    expect(screen.queryByText(TEXT)).toBeInTheDocument()
+    expect(screen.queryByText(FALLBACK)).not.toBeInTheDocument()
   })
 })
