+++ packages/react/src/hooks/useIsChanged.ts
@@ -1,0 +1,3 @@
+import { usePrevious } from '.'
+
+export const useIsChanged = (value: unknown) => usePrevious(value) !== value
