--- packages/react/src/hooks/useIsChanged/index.ts
@@ -1,1 +1,0 @@
-export { default as useIsChanged } from './useIsChanged'
