--- packages/react/src/hooks/useIsChanged/useIsChanged.ts
@@ -1,5 +1,0 @@
-import { usePrevious } from '../usePrevious'
-
-const useIsChanged = (value: unknown) => usePrevious(value) !== value
-
-export default useIsChanged
