+++ packages/react/src/hooks/useIsMounted.ts
@@ -1,6 +1,6 @@
 import { useEffect, useState } from 'react'
 
-const useIsMounted = () => {
+export const useIsMounted = () => {
   const [isMounted, setIsMounted] = useState(false)
 
   useEffect(() => {
@@ -9,5 +9,3 @@ const useIsMounted = () => {
 
   return isMounted
 }
-
-export default useIsMounted
