--- packages/react/src/hooks/useIsMounted/index.ts
@@ -1,1 +1,0 @@
-export { default as useIsMounted } from './useIsMounted'
