+++ packages/react/src/hooks/useKey.ts
@@ -1,10 +1,8 @@
 import { useCallback, useState } from 'react'
 
-const useKey = (): [number, () => void] => {
+export const useKey = (): [number, () => void] => {
   const [key, setKey] = useState(0)
   const refresh = useCallback(() => setKey((prev) => prev + 1), [])
 
   return [key, refresh]
 }
-
-export default useKey
