--- packages/react/src/hooks/useKey/index.ts
@@ -1,1 +1,0 @@
-export { default as useKey } from './useKey'
