--- packages/react/src/hooks/usePrevious/index.ts
@@ -1,1 +1,0 @@
-export { default as usePrevious } from './usePrevious'
