+++ packages/react/src/hooks/useRerender.ts
@@ -2,10 +2,8 @@ import { useCallback, useReducer } from 'react'
 
 type Rerender = () => void
 const reducer = (state: number) => state + 1
-const useRerender = (): Rerender => {
+export const useRerender = (): Rerender => {
   const [, dispatch] = useReducer(reducer, 0)
 
   return useCallback(() => dispatch(), [])
 }
-
-export default useRerender
