--- packages/react/src/hooks/useRerender/index.ts
@@ -1,1 +1,0 @@
-export { default as useRerender } from './useRerender'
