+++ packages/react/src/index.ts
@@ -1,14 +1,8 @@
 export { SuspensiveProvider, Suspensive } from './SuspensiveProvider'
-export { Suspense } from './Suspense'
-export { ErrorBoundary } from './ErrorBoundary'
-export { ErrorBoundaryGroup, useErrorBoundaryGroup } from './ErrorBoundaryGroup'
-export { AsyncBoundary } from './AsyncBoundary'
-export { Delay } from './Delay'
-
-export { withSuspense } from './withSuspense'
-export { withErrorBoundary } from './withErrorBoundary'
-export { withAsyncBoundary } from './withAsyncBoundary'
-export { withErrorBoundaryGroup } from './withErrorBoundaryGroup'
-export { withDelay } from './withDelay'
+export { Suspense, withSuspense } from './Suspense'
+export { ErrorBoundary, withErrorBoundary } from './ErrorBoundary'
+export { ErrorBoundaryGroup, withErrorBoundaryGroup, useErrorBoundaryGroup } from './ErrorBoundaryGroup'
+export { AsyncBoundary, withAsyncBoundary } from './AsyncBoundary'
+export { Delay, withDelay } from './Delay'
 
 export type { ErrorBoundaryFallbackProps } from './ErrorBoundary'
