+++ packages/react/src/types/PropsWithoutChildren.ts
@@ -1,0 +1,3 @@
+import { ComponentProps, ComponentType } from 'react'
+
+export type PropsWithoutChildren<TProps extends ComponentProps<ComponentType>> = Omit<TProps, 'children'>
