+++ packages/react/src/types/Tuple.ts
@@ -1,0 +1,1 @@
+export type Tuple<T = unknown> = T[] | readonly T[]
