+++ packages/react/src/utils/hasResetKeysChanged.ts
@@ -1,0 +1,2 @@
+export const hasResetKeysChanged = (a: unknown[] = [], b: unknown[] = []) =>
+  a.length !== b.length || a.some((item, index) => !Object.is(item, b[index]))
