--- packages/react/src/utils/hasResetKeysChanged/hasResetKeysChanged.ts
@@ -1,4 +1,0 @@
-const hasResetKeysChanged = (a: unknown[] = [], b: unknown[] = []) =>
-  a.length !== b.length || a.some((item, index) => !Object.is(item, b[index]))
-
-export default hasResetKeysChanged
