--- packages/react/src/utils/hasResetKeysChanged/index.ts
@@ -1,1 +1,0 @@
-export { default as hasResetKeysChanged } from './hasResetKeysChanged'
