--- packages/react/src/withSuspense.tsx
@@ -1,39 +1,0 @@
-import { ComponentType } from 'react'
-import { Suspense } from './Suspense'
-import { ComponentPropsWithoutChildren } from './types'
-
-export function withSuspense<Props extends Record<string, unknown> = Record<string, never>>(
-  Component: ComponentType<Props>,
-  suspenseProps?: ComponentPropsWithoutChildren<typeof Suspense>
-) {
-  const Wrapped = (props: Props) => (
-    <Suspense {...suspenseProps}>
-      <Component {...props} />
-    </Suspense>
-  )
-
-  if (process.env.NODE_ENV !== 'production') {
-    const name = Component.displayName || Component.name || 'Component'
-    Wrapped.displayName = `withSuspense(${name})`
-  }
-
-  return Wrapped
-}
-
-withSuspense.CSROnly = function withSuspenseCSROnly<Props extends Record<string, unknown> = Record<string, never>>(
-  Component: ComponentType<Props>,
-  suspenseProps?: ComponentPropsWithoutChildren<typeof Suspense.CSROnly>
-) {
-  const Wrapped = (props: Props) => (
-    <Suspense.CSROnly {...suspenseProps}>
-      <Component {...props} />
-    </Suspense.CSROnly>
-  )
-
-  if (process.env.NODE_ENV !== 'production') {
-    const name = Component.displayName || Component.name || 'Component'
-    Wrapped.displayName = `withSuspense.CSROnly(${name})`
-  }
-
-  return Wrapped
-}
